+++ src/lexy/lexy.py
@@ -3,8 +3,6 @@ from lexy_helpers.output_helpers import LexyFinder, LexyInit
 import click
 import json
 
-lexy_scraper = LexyScraper()
-
 
 @click.command()
 @click.version_option()
@@ -19,8 +17,12 @@ def lexy(language):
     - Use "update" to force update Lexy.
 
     - Use "modified" to view the last modified date of Lexy.
+
+    - Use "editor" to set up your default editor for viewing documentation.
     """
-    LexyInit(lexy_scraper).ensure_languages_file()
+    lexy_scraper = LexyScraper()
+    lexy_init = LexyInit(lexy_scraper)
+    lexy_init.ensure_languages_file()
     lexy_scraper.auto_update()
     language = language.lower()
     with open(f"{lexy_scraper.json_path}/languages.json", "r", encoding="utf-8") as f:
@@ -34,5 +36,11 @@ def lexy(language):
             case "modified":
                 last_modified = lexy_scraper.last_modified()
                 click.echo(f"The most recent update to Lexy was on: {last_modified}")
+            case "editor":
+                click.echo(
+                    f"\nYour current default editor is set to {lexy_finder.editor}\n"
+                )
+                if click.confirm("Do you want to change it?", abort=True):
+                    lexy_init.default_editor_setup()
             case _:
                 lexy_finder.language_finder(language)